// src/app/model/request/student-profile-request.ts
export interface StudentProfileRequest {
  studentId: number;
  addressId: number;
  studentFirstName: string;
  studentLastName: string;
  classStandard: string;
  gender: string;
  studentStatus: string;
  dateOfBirth: string;
  dateOfJoin: string;
  doorNum: string;
  street: string;
  addrLine: string;
  city: string;
  state: string;
  pincode: number;
}